refactor: structure 28-create-reducer-for-achive-and-save-kind-of-sort
