+++ src/NavigationPages/HistoryPage/helpers.ts
