+++ src/NavigationPages/StatisticsPage/index.tsx
